InputContainer를 생성하고 관심사를 분리하는 리팩터링
