+++ src/Page.test.jsx
@@ -8,27 +8,17 @@ jest.mock('react-redux');
 
 test('Page', () => {
   useSelector.mockImplementation((selector) => selector({
+    taskTitle: '',
     tasks: [
       { id: 1, title: 'Task-1' },
       { id: 2, title: 'Task-2' },
     ],
   }));
 
-  const handleChangeTitle = jest.fn();
-  const handleClickAddTask = jest.fn();
-
   const { getByText } = render((
-    <Page
-      taskTitle=""
-      onChangeTitle={handleChangeTitle}
-      onClickAddTask={handleClickAddTask}
-    />
+    <Page />
   ));
 
   expect(getByText(/Task-1/)).not.toBeNull();
   expect(getByText(/Task-2/)).not.toBeNull();
-
-  fireEvent.click(getByText('추가'));
-
-  expect(handleClickAddTask).toBeCalled();
 });
